@@ -1,6 +1,9 @@
 <?php
+declare(strict_types=1);
 
-namespace chilimatic\lib\Config;
+namespace chilimatic\lib\Config\Adapter;
+use chilimatic\lib\Config\Engine\DataStructure\Node;
+use chilimatic\lib\Config\IConfig;
 
 /**
  * Class AbstractConfig
@@ -43,9 +46,9 @@ abstract class AbstractConfig implements IConfig
         $this->mainNode = new Node(null, IConfig::MAIN_NODE_KEY, null);
 
         // add custom parameters
-        if (is_array($param) && count($param)) {
+        if ($param && is_array($param)) {
             // set the given parameters
-            foreach ($param as $key => $value) {
+            foreach ((array) $param as $key => $value) {
                 $node = new Node($this->mainNode, $key, $value, self::INIT_PARAMETER);
                 $this->mainNode->addChild($node);
             }
@@ -69,7 +72,7 @@ abstract class AbstractConfig implements IConfig
      *
      * @return mixed
      */
-    public function delete($key = '')
+    public function delete(string $key = '') : bool
     {
         $nodeList = $this->mainNode->getByKey($key);
         if (empty($nodeList)) {
@@ -90,7 +93,7 @@ abstract class AbstractConfig implements IConfig
      *
      * @return bool
      */
-    public function deleteById($id = '')
+    public function deleteById(string $id = '') : bool
     {
         $node = $this->mainNode->getById($id);
 
@@ -98,7 +101,7 @@ abstract class AbstractConfig implements IConfig
             true;
         }
 
-        return $node->delete();
+        return ($node->delete() ? true : false);
     }
 
     /**
@@ -108,7 +111,7 @@ abstract class AbstractConfig implements IConfig
      *
      * @return mixed
      */
-    public function get($var)
+    public function get(string $var)
     {
         if (!$this->mainNode) {
             return null;
@@ -130,7 +133,7 @@ abstract class AbstractConfig implements IConfig
      * @internal param $var
      * @return mixed
      */
-    public function getById($id)
+    public function getById(string $id)
     {
         $node = $this->mainNode->getById($id);
         if ($node === null) {
@@ -148,7 +151,7 @@ abstract class AbstractConfig implements IConfig
      *
      * @return mixed
      */
-    public function setById($id, $val)
+    public function setById(string $id, $val)
     {
         // set the variable
         if (empty($id)) {
@@ -171,19 +174,21 @@ abstract class AbstractConfig implements IConfig
      *
      * @return mixed
      */
-    public function set($key, $val)
+    public function set(string $key, $val)
     {
         // set the variable
         if (empty($key)) {
             return $this;
         }
 
-        if (!($node = $this->mainNode->getLastByKey($key))) {
+        $node = $this->mainNode->getLastByKey($key);
+
+        if (!$node || !$node->getParent()) {
             $newNode = new Node($this->mainNode, $key, $val);
             $this->mainNode->addChild($newNode);
         } else {
-            $newNode = new Node($node, $key, $val);
-            $node->addChild($newNode);
+            $newNode = new Node($node->getParent(), $key, $val);
+            $node->getParent()->addChild($newNode);
         }
 
         $this->lastNewNode = $newNode;
