@@ -9,14 +9,17 @@
  */
 
 namespace chilimatic\lib\Config;
+use chilimatic\lib\Config\Adapter\AbstractConfig;
 
 /**
  * Class ConfigFactory
  *
- * @package chilimatic\lib\config
+ * @package chilimatic\lib\Config
  */
 class ConfigFactory
 {
+
+    const ADAPTER_NAMESPACE = __NAMESPACE__ . '\\Adapter\\';
 
     /**
      * Factory for creating Config objects
@@ -30,13 +33,15 @@ class ConfigFactory
      *
      * @return AbstractConfig
      */
-    public static function make($type, $param = null)
+    public static function make(string $type, $param = null) : AbstractConfig
     {
         if (!$type) {
             throw new \LogicException("The Config Type has to be specified ... \$type is empty");
         }
 
-        $className = (string)__NAMESPACE__ . '\\' . (string)ucfirst($type);
+
+
+        $className = self::ADAPTER_NAMESPACE . (string)ucfirst($type);
 
         if (!class_exists($className, true)) {
             throw new \LogicException("The Config class has to be implemented and accessible ... $className is not found");
