@@ -1,10 +1,11 @@
 <?php
-namespace chilimatic\lib\config;
+namespace chilimatic\lib\Config;
+use chilimatic\lib\Config\Engine\DataStructure\Node;
 
 /**
  * Interface Config_Interface
  *
- * @package chilimatic\lib\config
+ * @package chilimatic\lib\Config
  */
 Interface IConfig
 {
@@ -74,7 +75,7 @@ Interface IConfig
      *
      * @return bool
      */
-    public function delete($key = '');
+    public function delete(string $key = '');
 
     /**
      * gets a specific parameter
@@ -83,7 +84,7 @@ Interface IConfig
      *
      * @return mixed
      */
-    public function get($key);
+    public function get(string $key);
 
 
     /**
@@ -93,7 +94,7 @@ Interface IConfig
      *
      * @return mixed
      */
-    public function getbyId($id);
+    public function getbyId(string $id);
 
     /**
      * gets a specific parameter
@@ -103,7 +104,7 @@ Interface IConfig
      *
      * @return mixed
      */
-    public function setById($id, $val);
+    public function setById(string $id, $val);
 
     /**
      * sets a specific parameter
@@ -113,7 +114,7 @@ Interface IConfig
      *
      * @return mixed
      */
-    public function set($key, $val);
+    public function set(string $key, $val);
 
     /**
      * saves the specified config
